// src/app/Models/produto.ts
export interface Produto{
    id?: number;
    titulo?: string;
    marca?: string;
    tamanho?: number
    preco?: number;
    descricao?: string;
    imagem?: string;
    status?: boolean;
    quantidade?: number;
    categoria?: string;
    data?: string;
    hex?: string;
    idStorage?: number;
    cor?: string;
    quantidadeComprada?: string;
}